fix: rework content config
