+++ src/pages/[locale]/blog/[slug].astro
@@ -1,7 +1,7 @@
 ---
 import { type CollectionEntry, getCollection } from "astro:content";
 import BlogPostPage from "@/components/pages/blog-post-page.astro";
-import type { BlogCollection } from "@/content/config";
+import type { BlogCollection } from "@/content.config";
 import { defaultLocale, locales } from "@/i18n/locales";
 import { getBlogEntry } from "@/i18n/utils";
 
@@ -19,7 +19,7 @@ export async function getStaticPaths() {
     const posts = await getCollection(blogEntry);
     for (const p of posts) {
       paths.push({
-        params: { slug: p.slug, locale: l },
+        params: { slug: p.id, locale: l },
         props: { post: p },
       });
     }
