+++ src/pages/blog/[slug].astro
@@ -1,7 +1,7 @@
 ---
 import { type CollectionEntry, getCollection } from "astro:content";
 import BlogPostPage from "@/components/pages/blog-post-page.astro";
-import type { BlogCollection } from "@/content/config";
+import type { BlogCollection } from "@/content.config";
 import { defaultLocale } from "@/i18n/locales";
 import { getBlogEntry } from "@/i18n/utils";
 
@@ -9,7 +9,7 @@ export async function getStaticPaths() {
   const blogEntry = getBlogEntry(defaultLocale);
   const posts = await getCollection(blogEntry);
   return posts.map((p) => ({
-    params: { slug: p.slug },
+    params: { slug: p.id },
     props: { post: p },
   }));
 }
